13% faster on penryn, 16% on sandybridge, 15% on bulldozer Not simd; a compiler should have generated this, but gcc didn't.
--
I don't understand where the bottleneck is.
This asm version is 10 cycles per pixel on sandybridge.
The loop is 23 instructions (all single-uop). 6 loads, 3 stores, 13
arithmetic, 1 jump.
If it were latency-bound (by the horizontal filter's dependency on the
previous iteration), it should take 8 cycles.
If it were throughput bound, it should take 6 cycles.
It shouldn't be decode-bound, since the loop easily fits in the uop cache.
If I remove everything except the horizontal filter, it does indeed take 8
cycles.
If I keep all of the instructions but remove the data-dependency of LUT
addresses, it takes 7 cycles.
So neither latency nor throughput explains the observed speed.
And it can't be due to limited OOE, because manually interleaving two
iterations doesn't help.